played around with expanding other functions
The way I'm doing this is super inefficient. It also doesn't jive well
with the stuff I set up for mkf.delta(). I need to modify mkgif() to be
a little more general, so I can use it with other expansions. I also
need to "bake" the expansions so they aren't re-done for each
iteration. (I think they may even be re-done for each point right now,
lol...) It might be cute to make a bake() function, so I don't have to
do it explicitly each time, and I can still use the functional goodness
I make for expand() unexpand() and approximate().